Extract ServiceCreator out to its own class
